# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Courtside TW API
  version: 1.0.0
  description: Contract-first REST API for public reading, reader library, editorial publication and optional provenance/SIWE boundaries.
servers:
  - url: https://api.courtside.tw
    description: Production API origin placeholder; deployment is a later task.
...
security:
  - bffSession: []
  - oidcBearer: []
paths:
...
  /api/v1/editor/audit:
    get:
      tags:
        - Editorial
      operationId: listEditorialAudit
      summary: List append-only editorial audit events
      description: Reads immutable audit events for one article, issue, or media asset.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - name: targetType
          in: query
          required: true
          schema:
            type: string
            enum:
              - ARTICLE
              - ISSUE
              - MEDIA_ASSET
        - name: targetId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Audit events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventPage"
              examples:
                valid:
                  value:
                    items: []
                    page:
                      nextCursor: null
                      limit: 50
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "429":
          $ref: "#/components/responses/Problem429"
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC issuer/audience validated JWT for Spring Security.
    bffSession:
      type: apiKey
      in: cookie
      name: __Host-courtside_session
      description: Secure, HttpOnly, SameSite BFF session cookie.
  parameters:
...
    Limit:
      name: limit
      in: query
      required: false
      description: Bounded page size.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  headers:
    XRequestId:
      description: Trace/request identifier. Never contains tokens or article content.
      schema:
        type: string
        minLength: 1
        maxLength: 100
...
    RetryAfter:
      description: Seconds until the caller may retry.
      schema:
        type: integer
        minimum: 1
        maximum: 3600
  schemas:
    Uuid:
      type: string
      format: uuid
      description: Stable public identifier.
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
        - requestId
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
          maxLength: 1000
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          minLength: 1
          maxLength: 100
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - AUTHENTICATION_REQUIRED
            - FORBIDDEN
            - RESOURCE_NOT_FOUND
            - VERSION_CONFLICT
            - RIGHTS_OR_CONTENT_GATE
            - RATE_LIMITED
        errors:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      additionalProperties: false
      required:
        - path
        - code
        - message
      properties:
        path:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          minLength: 1
          maxLength: 80
        message:
          type: string
          minLength: 1
          maxLength: 500
    PageMeta:
      type: object
      additionalProperties: false
      required:
        - nextCursor
        - limit
      properties:
        nextCursor:
          type:
            - string
            - "null"
          maxLength: 256
        limit:
          type: integer
          minimum: 1
          maximum: 100
    AuditEvent:
      type: object
      additionalProperties: false
      required:
        - id
        - occurredAt
        - actorSubject
        - action
        - targetType
        - targetId
        - metadata
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        occurredAt:
          type: string
          format: date-time
        actorSubject:
          type: string
        action:
          type: string
        targetType:
          type: string
        targetId:
          $ref: "#/components/schemas/Uuid"
        requestId:
          type: [string, "null"]
        metadata:
          type: object
          additionalProperties: true
    AuditEventPage:
      type: object
      additionalProperties: false
      required:
        - items
        - page
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
        page:
          $ref: "#/components/schemas/PageMeta"
...
  responses:
    Problem400:
      description: Invalid request (400).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: INVALID_REQUEST example
              value:
                type: https://courtside.tw/problems/invalid_request
                title: Invalid request
                status: 400
                detail: The request is invalid.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: INVALID_REQUEST
      x-stable-error-code: INVALID_REQUEST
    Problem401:
      description: Authentication required (401).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: AUTHENTICATION_REQUIRED example
              value:
                type: https://courtside.tw/problems/authentication_required
                title: Authentication required
                status: 401
                detail: Authentication is required.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: AUTHENTICATION_REQUIRED
      x-stable-error-code: AUTHENTICATION_REQUIRED
    Problem403:
      description: Forbidden (403).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: FORBIDDEN example
              value:
                type: https://courtside.tw/problems/forbidden
                title: Forbidden
                status: 403
                detail: The actor is not allowed to perform this operation.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: FORBIDDEN
      x-stable-error-code: FORBIDDEN
...
    Problem429:
      description: Too many requests (429).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: RATE_LIMITED example
              value:
                type: https://courtside.tw/problems/rate_limited
                title: Too many requests
                status: 429
                detail: The request rate is limited.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: RATE_LIMITED
      x-stable-error-code: RATE_LIMITED
...